# docs/swagger.yaml
definitions:
  domain.DepositRequest:
    properties:
      amount:
        type: string
      callback_url:
        type: string
    required:
    - amount
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.TransferRequest:
    properties:
      amount:
        type: string
      callback_url:
        type: string
      to:
        type: string
    required:
    - amount
    - to
    type: object
  domain.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.WithdrawRequest:
    properties:
      amount:
        type: string
      callback_url:
        type: string
    required:
    - amount
    type: object
info:
  contact: {}
  description: API do FinancialSystemPro
  title: FinancialSystemPro API
  version: "1.0"
paths:
  /api/balance:
    get:
      consumes:
      - application/json
      description: Endpoint para consultar o saldo da conta do usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Consulta o saldo da conta do usuário
      tags:
      - Transactions
  /api/deposit:
    post:
      consumes:
      - application/json
      description: Endpoint para depositar valor na conta do usuário
      parameters:
      - description: Dados do depósito
        in: body
        name: depositRequest
        required: true
        schema:
          $ref: '#/definitions/domain.DepositRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deposita valor na conta do usuário
      tags:
      - Transactions
  /api/login:
    post:
      consumes:
      - application/json
      description: Endpoint para autenticar usuário
      parameters:
      - description: Dados de login
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Autentica usuário
      tags:
      - Users
  /api/transfer:
    post:
      consumes:
      - application/json
      description: Endpoint para transferir valor para outra conta
      parameters:
      - description: Dados da transferência
        in: body
        name: transferRequest
        required: true
        schema:
          $ref: '#/definitions/domain.TransferRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Transfere valor para outra conta
      tags:
      - Transactions
  /api/users:
    post:
      consumes:
      - application/json
      description: Endpoint para criar usuário
      parameters:
      - description: Dados do usuário
        in: body
        name: userRequest
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria um novo usuário
      tags:
      - Users
  /api/withdraw:
    post:
      consumes:
      - application/json
      description: Endpoint para retirar valor da conta do usuário
      parameters:
      - description: Dados do saque
        in: body
        name: withdrawRequest
        required: true
        schema:
          $ref: '#/definitions/domain.WithdrawRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Retira valor da conta do usuário
      tags:
      - Transactions
securityDefinitions:
  BearerAuth:
    description: Digite "Bearer {seu_token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"