# domain/repositories.go
package domain

type UserRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

type LoginRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

type DepositRequest struct {
	Amount      string `json:"amount" validate:"required"`
	CallbackURL string `json:"callback_url" validate:"omitempty,url"`
}

type BalanceRequest struct {
	UserID string `json:"user_id" validate:"required"`
}

type WithdrawRequest struct {
	Amount      string `json:"amount" validate:"required"`
	CallbackURL string `json:"callback_url" validate:"omitempty,url"`
}

type TransferRequest struct {
	Amount      string `json:"amount" validate:"required"`
	To          string `json:"to" validate:"required,email"`
	CallbackURL string `json:"callback_url" validate:"omitempty,url"`
}